Commit 4. Cambios en el tipo de dato de "price" y "percentage" en las clases "Purchase", "PurchasingManager", "Product" y "Discount. Mejora en la visualización del output de la App

+++ src/org/ulpgc/is1/model/Purchase.java
@@ -40,7 +40,7 @@ public class Purchase {
         return deliveryAddress;
     }
 
-    public int price() {
+    public float price() {
         return product.getPrice();
     }
